taa_django/taa_app/views.py
from django.shortcuts import render
from django.http import HttpResponseRedirect
from django.contrib.auth.models import User
from rest_framework import permissions, status
from rest_framework.decorators import api_view
from rest_framework.response import Response
from rest_framework.views import APIView
from .serializers import UserSerializer, UserSerializerWithToken, HikeSerializer, BrewerySerializer
from rest_framework import viewsets
from . import models



@api_view(['GET'])
def current_user(request):
    serializer = UserSerializer(request.user)
    return Response(serializer.data)


class UserList(APIView):
    permission_classes = (permissions.AllowAny,)

    def post(self, request, format=None):
        serializer = UserSerializerWithToken(data=request.data)
        if serializer.is_valid():
            serializer.save()
            return Response(serializer.data, status=status.HTTP_201_CREATED)
        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)

class HikeViewSet(viewsets.ModelViewSet):
    query = models.Hike.objects.all()
    serializer = HikeSerializer

class BreweryViewSet(viewsets.ModelViewSet):
    query = models.Brewery.objects.all()
    serializer = BrewerySerializer 